+++ packages/sources/src/components/form/models/form-group.model.ts
@@ -1,3 +1,4 @@
+import { FormGroupCheckboxes } from './form-group-checkboxes.model';
 import { FormGroupInput, FormGroupInputDate } from './form-group-input.model';
 
-export type FormGroup<TemplateFnReturnType> = FormGroupInput<TemplateFnReturnType> | FormGroupInputDate<TemplateFnReturnType>;
+export type FormGroup<TemplateFnReturnType> = FormGroupCheckboxes<TemplateFnReturnType> | FormGroupInput<TemplateFnReturnType> | FormGroupInputDate<TemplateFnReturnType>;