+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 * **core:** dso-ozon-content met `[interactive]` als interactief element visualiseren + click event emitten ([#1569](https://github.com/dso-toolkit/dso-toolkit/issues/1569))
 
 ### Changed
+* **dso-toolkit + css:** Form: "Checkboxes" verhuizen naar @dso-toolkit/css + Storybook ([#1559](https://github.com/dso-toolkit/dso-toolkit/issues/1559))
 * **dso-toolkit + css:** Overblijfselen "Search Bar" verhuizen naar `@dso-toolkit/css` ([#1466](https://github.com/dso-toolkit/dso-toolkit/issues/1466))
 * **dso-toolkit:** NPM dependency `jquery-pjax` van een versie voorzien ([#1594](https://github.com/dso-toolkit/dso-toolkit/issues/1594))